fixed server paths
