+++ package.json
@@ -5,7 +5,8 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "eject": "expo eject"
+    "eject": "expo eject",
+    "server": "node server/server.js"
   },
   "dependencies": {
     "@react-native-community/geolocation": "^2.0.2",
